polish chamber experiment delivery
