+++ app/directions/chambers/chamber-experiment.module.css
@@ -355,7 +355,7 @@
   }
 
   .identity {
-    top: 54%;
+    top: 46%;
     width: calc(100% - 1.5rem);
   }
